Create test.sh
Bash script to test the client & server.
The script creates two clients for the same user and makes changes in both of their directories.

The pull_rate parameter is set to 1 second. Therefore, within a maximum of 1 second, all the directories (user1client1, user1client2, and AllClients/[user_id]) should be identical.